+++ src/lib/utils/gemini.util.ts
@@ -3,15 +3,15 @@ import { GoogleGenAI } from '@google/genai';
 import { ENV, ENV_ERROR } from '@/constants/env.constant';
 import { FoodAnalysisResult } from '@/types/gemini.type';
 
-export function getGenerativeAI() {
+export const getGenerativeAI = () => {
   if (!ENV.GEMINI_KEY) throw new Error(ENV_ERROR.GEMINI_KEY);
   return new GoogleGenerativeAI(ENV.GEMINI_KEY);
-}
+};
 
-export function getGenAI() {
+export const getGenAI = () => {
   if (!ENV.GEMINI_KEY) throw new Error(ENV_ERROR.GEMINI_KEY);
   return new GoogleGenAI({ apiKey: ENV.GEMINI_KEY });
-}
+};
 
 /**
  * Gemini 응답 문자열을 안전하게 파싱하여 JSON 배열로 변환
@@ -21,7 +21,7 @@ export function getGenAI() {
  * @param {string} text - Gemini JSON 형식 응답 문자열
  * @returns {FoodAnalysisResult[]} - JSON으로 변환된 응답의 Array
  */
-export function parseGeminiResponse(text: string): FoodAnalysisResult[] {
+export const parseGeminiResponse = (text: string): FoodAnalysisResult[] => {
   try {
     const cleaned = text.replace(/```json|```/g, '').trim();
     const parsed = JSON.parse(cleaned);
@@ -35,4 +35,4 @@ export function parseGeminiResponse(text: string): FoodAnalysisResult[] {
     console.error('Failed to parse Gemini response:', parseError);
     return [];
   }
-}
+};
